- create queue for some cmd packet readers - return dead bush in deserts - dead bush now drop sticks 0...2 - return block destroy particles - bugix increment method in player inventory

+++ node_server/network/clientpackets/cmd_drop_item_pickup.js
@@ -1,3 +1,4 @@
+import { PickatActions } from "../../../www/js/block_action.js";
 import { ServerClient } from "../../../www/js/server_client.js";
 import { PlayerEvent } from "../../player_event.js";
 
@@ -29,9 +30,12 @@ export default class packet_reader {
             if(!drop_item) {
                 continue;
             }
+
             // check dist
             const dist = drop_item.pos.distance(player.state.pos);
             if(dist > MAX_DIST_FOR_PICKUP) {
+                console.error(`ERROR: pickup item so far from player ${dist}m > ${MAX_DIST_FOR_PICKUP}`);
+                this.restoreDropItemForPlayer(player, drop_item);
                 continue;
             }
 
@@ -43,24 +47,41 @@ export default class packet_reader {
 
             // 1. add items to inventory
             const items = drop_item.items;
+            const restored_items = [];
             for(const item of items) {
-                const ok = player.inventory.increment(item);
-                // @todo check if not ok!
+                const ok = player.inventory.increment(item, true);
+                // check if not ok!
+                if(!ok) {
+                    restored_items.push(item);
+                    continue;
+                }
+            }
+
+            // restore if not applyed no one item
+            if(restored_items.length == items.length) {
+                console.log(`restore if not applyed no one item ${restored_items.length} == ${items.length}`)
+                this.restoreDropItemForPlayer(player, drop_item);
+                continue;
             }
 
             // delete from chunk
             chunk.drop_items.delete(entity_id);
-
             // unload drop item
             drop_item.onUnload();
-
             // deactive drop item in database
             world.db.deleteDropItem(entity_id);
             
             // @todo players must receive this packet after 200ms after player send request
             // because animation not ended
-            setTimeout(() => {
-                
+            setTimeout(async () => {
+
+                // Create new drop item
+                if(restored_items.length > 0) {
+                    const actions = new PickatActions();
+                    actions.addDropItem({pos: drop_item.pos, items: restored_items, force: true});
+                    await world.applyActions(null, actions); 
+                }
+
                 // play sound on client
                 let packets_sound = [{
                     name: ServerClient.CMD_PLAY_SOUND,
@@ -83,6 +104,19 @@ export default class packet_reader {
 
         }
 
+        return true;
+
+    }
+
+    //
+    static restoreDropItemForPlayer(player, drop_item) {
+        // need to restore item in player game
+        const packets = [{
+            name: ServerClient.CMD_DROP_ITEM_ADDED,
+            data: [drop_item]
+        }];
+        player.sendPackets(packets);
+        return true;
     }
 
 }
